App as root, calling navigation for correctly wrapping redux
